Commit: Files were renamed in accordance with the norms; A common configuration file has been created for the database and the data-source file; The modules of films and genres have been renamed.
