+++ src/genre/favorite-genre.module.ts
@@ -1,10 +1,10 @@
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Module } from '@nestjs/common';
-import { FavoriteGenre } from './favoriteGenre.entity';
+import { FavoriteGenre } from './favorite-genre.entity';
 
 @Module({
   controllers: [],
   providers: [],
   imports: [TypeOrmModule.forFeature([FavoriteGenre])],
 })
-export class UserFavoriteGenreModule {}
+export class FavoriteGenreModule {}
