+++ src/movie/favorite-movie.entity.ts
