+++ src/movie/favorite-movie.module.ts
@@ -1,10 +1,10 @@
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Module } from '@nestjs/common';
-import { FavoriteMovie } from './favoriteMovie.entity';
+import { FavoriteMovie } from './favorite-movie.entity';
 
 @Module({
   controllers: [],
   providers: [],
   imports: [TypeOrmModule.forFeature([FavoriteMovie])],
 })
-export class UserFavoriteMovieModule {}
+export class FavoriteMovieModule {}
